BUG: fix the issue with warning popup due to invalid DB path
The warning popup reported in #35 was caused by setting a DICOM DB path that does
not have a DB (and was not supposed to be a real DB path but only to trigger update
of the DB view). The explicit trigger of database update is not necessary. This
commit removes the unnecessary DB reset and some additional unnecessary code.
Should resolve #35.